services/oracle/pyproject.toml
[project]
name = "nps-oracle"
version = "4.0.0"
description = "NPS V4 Oracle Service — FC60 numerology, readings, timing, AI analysis"
requires-python = ">=3.11"
dependencies = [
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "grpcio-health-checking>=1.60.0",
    "protobuf>=4.25.0",
    "anthropic>=0.39.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "ruff>=0.2.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100
